requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
pymupdf>=1.25.0
requests>=2.32.0
beautifulsoup4>=4.12.0
pydantic>=2.10.0
python-dotenv>=1.0.0
validators>=0.34.0
tenacity>=9.0.0
